% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_7davs_and_target}
\alias{add_7davs_and_target}
\title{Add 7dav and target to the data
Target is the updates made ref_lag days after the first release}
\usage{
add_7davs_and_target(df, value_col, refd_col, lag_col, ref_lag)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location
reported for each reference date and issue date.}

\item{value_col}{string specifying name of value (counts) field within
the input dataframe.}

\item{refd_col}{string specifying name of reference date field within
the input dataframe.}

\item{lag_col}{string specifying name of lag field within
the input dataframe.}

\item{ref_lag}{max lag to use for training}
}
\description{
Add 7dav and target to the data
Target is the updates made ref_lag days after the first release
}
