% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{add_lag}
\alias{add_lag}
\title{Add `lag` field based on issue date and time value, if needed}
\usage{
add_lag(df, lag_col = "lag", issued_col = "issue_date")
}
\arguments{
\item{df}{dataframe containing input data from a single parquet file}

\item{lag_col}{string specifying name of lag field within
the input dataframe.}

\item{issued_col}{string specifying name of issue date (version) field within
the input dataframe.}
}
\description{
Add `lag` field based on issue date and time value, if needed
}
