% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_sqrtscale}
\alias{add_sqrtscale}
\title{Add columns to indicate the scale of value at square root level}
\usage{
add_sqrtscale(train_data, test_data, max_raw, value_col)
}
\arguments{
\item{train_data}{Data Frame for training}

\item{test_data}{Data Frame for testing}

\item{value_col}{the column name of the considered value}

\item{the}{maximum value in the training data at square root level}
}
\description{
Add columns to indicate the scale of value at square root level
}
