% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_weekofmonth}
\alias{add_weekofmonth}
\title{Add one hot encoding for week of a month info in terms of issue date}
\usage{
add_weekofmonth(df, wm = week_issues, time_col)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location 
reported for each reference date and issue date.}

\item{wm}{vector of weeks of a month}

\item{time_col}{column used for the date, can be either reference date or 
issue date}
}
\description{
Add one hot encoding for week of a month info in terms of issue date
}
