% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir_not_exist}
\alias{create_dir_not_exist}
\title{Create directory if not already existing}
\usage{
create_dir_not_exist(path)
}
\arguments{
\item{path}{character vector giving the directory to create}
}
\description{
Create directory if not already existing
}
