% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{data_filteration}
\alias{data_filteration}
\title{Filtration for training and testing data with different lags}
\usage{
data_filteration(test_lag, geo_train_data, geo_test_data, lag_pad)
}
\arguments{
\item{test_lag}{integer number of days ago to predict for}

\item{geo_train_data}{training data for a certain location}

\item{geo_test_data}{testing data for a certain location}
}
\description{
Filtration for training and testing data with different lags
}
