% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{datetime_to_date}
\alias{datetime_to_date}
\title{Extract date from datetime column in timezone-aware way}
\usage{
datetime_to_date(df, cols = c("time_value", "issue_date"))
}
\arguments{
\item{df}{dataframe containing input data from a single parquet file}

\item{cols}{vector of names of datetime fields to convert to date}
}
\description{
Extract date from datetime column in timezone-aware way
}
