% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior_estimation.R
\name{delta}
\alias{delta}
\title{Functions for Beta Prior Approach.
This is used only for the ratio prediction e.g. fraction of Covid claims, 
percentage of positive tests. We assume that the ratio follows a beta distribution
that is day-of-week dependent. A quantile regression model is used first with lasso
penalty for supporting quantile estimation and then a non-linear minimization is used
for prior estimation. 
Sum of squared error}
\usage{
delta(fit, actual)
}
\arguments{
\item{fit}{estimated values}

\item{actual}{actual values}
}
\description{
Functions for Beta Prior Approach.
This is used only for the ratio prediction e.g. fraction of Covid claims, 
percentage of positive tests. We assume that the ratio follows a beta distribution
that is day-of-week dependent. A quantile regression model is used first with lasso
penalty for supporting quantile estimation and then a non-linear minimization is used
for prior estimation. 
Sum of squared error
}
