% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior_estimation.R
\name{est_priors}
\alias{est_priors}
\title{Main function for the beta prior approach
Estimate the priors for the beta distribution based on data for 
a certain day of a week}
\usage{
est_priors(
  train_data,
  prior_test_data,
  cov,
  taus,
  params_list,
  response,
  lp_solver,
  lambda,
  start = c(0, log(10)),
  base_pseudo_denom = 1000,
  base_pseudo_num = 10
)
}
\arguments{
\item{train_data}{Data Frame for training}

\item{prior_test_data}{Data Frame for testing}

\item{taus}{vector of considered quantiles}

\item{params_list}{the list of parameters for training}

\item{response}{the column name of the response variable}

\item{lp_solver}{the lp solver used in Quantgen}

\item{start}{the initialization of the the points in nlm}

\item{base_pseudo_num}{the pseudo counts added to numerator if little data for training}

\item{dw}{column name to indicate which day of a week it is}

\item{labmda}{the level of lasso penalty}

\item{base_pseudo_denum}{the pseudo counts added to denominator if little data for training}
}
\description{
Main function for the beta prior approach
Estimate the priors for the beta distribution based on data for 
a certain day of a week
}
