% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{evl}
\alias{evl}
\title{Evaluation of the test results based on WIS score
The WIS score calculation is based on the weighted_interval_score function 
from the `evalcast` package from Delphi}
\usage{
evl(test_data, taus)
}
\arguments{
\item{test_data}{multiple columns for the prediction results of different 
quantiles. Each row represents an update with certain (reference_date, 
issue_date, location) combination.}

\item{taus}{vector of quantiles interested}
}
\description{
Evaluation of the test results based on WIS score
The WIS score calculation is based on the weighted_interval_score function 
from the `evalcast` package from Delphi
}
