% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{exponentiate_preds}
\alias{exponentiate_preds}
\title{Un-log predicted values}
\usage{
exponentiate_preds(test_data, taus)
}
\arguments{
\item{test_data}{dataframe with a column containing the prediction results of
each requested quantile. Each row represents an update with certain
(reference_date, issue_date, location) combination.}

\item{taus}{numeric vector of quantiles to be predicted. Values
must be between 0 and 1.}
}
\description{
Un-log predicted values
}
