% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_test_result}
\alias{export_test_result}
\title{Export the result to customized directory}
\usage{
export_test_result(
  test_data,
  coef_data,
  indicator,
  signal,
  geo_level,
  signal_suffix,
  lambda,
  training_end_date,
  value_type,
  export_dir
)
}
\arguments{
\item{test_data}{test data containing prediction results}

\item{coef_data}{data frame containing the estimated coefficients}

\item{value_type}{string describing signal type. Either "count" or "fraction".}

\item{export_dir}{path to directory to save output to}

\item{model_path_prefix}{string containing necessary information}
}
\description{
Export the result to customized directory
}
