% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fill_rows}
\alias{fill_rows}
\title{Re-index, fill na, make sure all reference date have enough rows for updates}
\usage{
fill_rows(df, refd_col, lag_col, min_refd, max_refd, ref_lag)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location
reported for each reference date and issue date.}

\item{refd_col}{string specifying name of reference date field within
the input dataframe.}

\item{lag_col}{string specifying name of lag field within
the input dataframe.}

\item{min_refd}{the earliest reference date considered in the data}

\item{max_refd}{the latest reference date considered in the data}

\item{ref_lag}{max lag to use for training}
}
\value{
df_new Data Frame with filled rows for missing lags
}
\description{
Re-index, fill na, make sure all reference date have enough rows for updates
}
