% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_counties}
\alias{filter_counties}
\title{Subset list of counties to those included in the 200 most populous in the US}
\usage{
filter_counties(geos)
}
\arguments{
\item{geos}{character vector of county FIPS codes}
}
\description{
Subset list of counties to those included in the 200 most populous in the US
}
