% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fips_to_geovalue}
\alias{fips_to_geovalue}
\title{Make sure data contains a `geo_value` field}
\usage{
fips_to_geovalue(df)
}
\arguments{
\item{df}{Data Frame of aggregated counts within a single location
reported for each reference date and issue date.}
}
\description{
Make sure data contains a `geo_value` field
}
