% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_filename}
\alias{generate_filename}
\title{Construct filename for model with given parameters}
\usage{
generate_filename(
  indicator,
  signal,
  geo_level,
  signal_suffix,
  lambda,
  training_end_date = "",
  geo = "",
  value_type = "",
  test_lag = "",
  tau = "",
  dw = "",
  beta_prior_mode = FALSE,
  model_mode = TRUE
)
}
\arguments{
\item{indicator}{string specifying the name of the indicator as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{signal}{string specifying the name of the signal as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{signal_suffix}{string specifying value column name
ending to be appended to standard value column names from
`params$num_col` and `params$denom_col`. Used for non-standard
value column names and when processing multiple signals from a
single input dataframe, as with `quidel`'s age buckets.}

\item{lambda}{the level of lasso penalty}

\item{geo}{string specifying the name of the geo region (e.g. FIPS
code for counties)}

\item{value_type}{string describing signal type. Either "count" or "fraction".}

\item{test_lag}{integer number of days ago to predict for}

\item{tau}{decimal quantile to be predicted. Values must be between 0 and 1.}
}
\value{
path to file containing model object
}
\description{
Construct filename for model with given parameters
}
