% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_model_filename}
\alias{generate_model_filename}
\title{Construct filename for model with given parameters}
\usage{
generate_model_filename(
  indicator,
  signal,
  geo,
  signal_suffix,
  value_type,
  test_lag,
  tau,
  lambda
)
}
\arguments{
\item{indicator}{string specifying the name of the indicator as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{signal}{string specifying the name of the signal as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{geo}{string specifying the name of the geo region (e.g. FIPS
code for counties)}

\item{signal_suffix}{string specifying value column name
ending to be appended to standard value column names from
`params$num_col` and `params$denom_col`. Used for non-standard
value column names and when processing multiple signals from a
single input dataframe, as with `quidel`'s age buckets.}

\item{value_type}{string describing signal type. Either "count" or "fraction".}

\item{test_lag}{integer number of days ago to predict for}

\item{tau}{decimal quantile to be predicted. Values must be between 0 and 1.}

\item{lambda}{the level of lasso penalty}
}
\value{
path to file containing model object
}
\description{
Construct filename for model with given parameters
}
