% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_7dav}
\alias{get_7dav}
\title{Calculate 7 day moving average for each issue date}
\usage{
get_7dav(pivot_df, refd_col)
}
\arguments{
\item{pivot_df}{Data Frame where the columns are issue dates and the rows are 
reference dates}

\item{refd_col}{string specifying name of reference date field within
the input dataframe.}
}
\description{
The 7dav for date D reported on issue date D_i is the average from D-7 to D-1
}
