% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_issue_date_range}
\alias{get_issue_date_range}
\title{Find the earliest and latest issue dates needed for either training or testing.}
\usage{
get_issue_date_range(params)
}
\arguments{
\item{params}{named list containing modeling and data settings. Must include
the following elements: `ref_lag`, `testing_window`, `test_dates`,
`training_days`, `num_col`, `denom_col`, `taus`, `lambda`, `export_dir`,
`lp_solver`, `input_dir`, `cache_dir`, `geo_levels`, and `value_types`.}
}
\description{
With current logic, we always need to include data for model training (in
case cached models are not available for a "make_predictions"-only run and
we need to train new models).
}
\details{
We generate test and train data by applying the following filters:
  - Test data is data where issue_date is in params$test_dates
    (as a continuous filter, min(params$test_dates) <= issue_date <= max(params$test_dates) )
  - Train data is data where issue_date < training_end_date; and
    training_start_date < target_date <= training_end_date

Train data doesn't have an explicit lower bound on issue_date, but we can
derive one.

Since target_date = reference_date + params$ref_lag and issue_date >=
reference_date, the requirement that training_start_date < target_date
also implies that issue date must be > training_start_date - params$ref_lag
}
