% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_model}
\alias{get_model}
\title{Train model using quantile regression with Lasso penalty, or load from disk}
\usage{
get_model(
  model_path,
  train_data,
  covariates,
  tau,
  lambda,
  lp_solver,
  train_models
)
}
\arguments{
\item{model_path}{path to read model from or to save model to}

\item{train_data}{Data Frame containing training data}

\item{covariates}{character vector of column names serving as the covariates for the model}

\item{tau}{decimal quantile to be predicted. Values must be between 0 and 1.}

\item{lambda}{the level of lasso penalty}

\item{lp_solver}{string specifying the lp solver to use in
Quantgen fitting. Either "glpk" or "gurobi". For faster
optimization, use Gurobi (requires separate installation
of the `gurobi` package).}

\item{train_models}{boolean indicating whether to train models (TRUE). If
FALSE previously trained models (stored locally) will be used instead.
Default is TRUE.}
}
\description{
Train model using quantile regression with Lasso penalty, or load from disk
}
