% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_key}
\alias{make_key}
\title{Generate key for identifying a value_type-signal combo}
\usage{
make_key(value_type, signal_suffix)
}
\arguments{
\item{value_type}{string describing signal type. Either "count" or "fraction".}

\item{signal_suffix}{string specifying value column name
ending to be appended to standard value column names from
`params$num_col` and `params$denom_col`. Used for non-standard
value column names and when processing multiple signals from a
single input dataframe, as with `quidel`'s age buckets.}
}
\description{
Generate key for identifying a value_type-signal combo
}
