% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_training_and_testing}
\alias{model_training_and_testing}
\title{Model training and prediction using quantile regression with Lasso penalty
The quantile regression uses the quantile_lasso function from quantgen package}
\usage{
model_training_and_testing(
  train_data,
  test_data,
  taus,
  covariates,
  lp_solver,
  lambda,
  test_date,
  geo
)
}
\arguments{
\item{train_data}{Data frame for training}

\item{test_data}{Data frame for testing}

\item{taus}{numeric vector of quantiles to be predicted. Values
must be between 0 and 1.}

\item{covariates}{list of column names serving as the covariates for the model}

\item{lp_solver}{the lp solver used in Quantgen}

\item{lambda}{the level of lasso penalty}

\item{test_date}{Date object representing test date}

\item{geo}{string specifying the name of the geo region (e.g. FIPS
code for counties)}
}
\description{
Model training and prediction using quantile regression with Lasso penalty
The quantile regression uses the quantile_lasso function from quantgen package
}
