% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{moving_average}
\alias{moving_average}
\title{Right-aligned moving average to replace `zoo::rollmeanr`}
\usage{
moving_average(x, n)
}
\arguments{
\item{x}{numeric vector}

\item{n}{integer indicating how many total points (current included) to use
in mean}
}
\description{
Right-aligned moving average to replace `zoo::rollmeanr`
}
