% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_ts}
\alias{msg_ts}
\title{Write a message to the console with the current time}
\usage{
msg_ts(...)
}
\arguments{
\item{...}{the body of the message to display. Objects should be strings or
coercible to strings and are pasted together with no separator.}
}
\description{
Write a message to the console with the current time
}
