% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior_estimation.R
\name{objective}
\alias{objective}
\title{Generate objection function}
\usage{
objective(theta, x, prob, ...)
}
\arguments{
\item{theta}{parameters for the distribution in log scale}

\item{prob}{the expected probabilities}
}
\description{
Generate objection function
}
