% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{params_element_exists_and_valid}
\alias{params_element_exists_and_valid}
\title{Check if an element in params exists and is not missing}
\usage{
params_element_exists_and_valid(params, key)
}
\arguments{
\item{params}{named list containing modeling and data settings. Must include
the following elements: `ref_lag`, `testing_window`, `test_dates`,
`training_days`, `num_col`, `denom_col`, `taus`, `lambda`, `export_dir`,
`lp_solver`, `input_dir`, `cache_dir`, `geo_levels`, and `value_types`.}

\item{key}{string indicating name of element within `params` to check}
}
\description{
Check if an element in params exists and is not missing
}
