% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior_estimation.R
\name{ratio_adj}
\alias{ratio_adj}
\title{Update ratio using beta prior approach}
\usage{
ratio_adj(train_data, test_data, prior_test_data)
}
\arguments{
\item{train_data}{training data}

\item{test_data}{testing data}

\item{prior_test_data}{testing data for the lag -1 model}
}
\description{
Update ratio using beta prior approach
}
