% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_data}
\alias{read_data}
\title{Read a parquet file into a dataframe}
\usage{
read_data(input_dir)
}
\arguments{
\item{input_dir}{path to the directory containing input data}
}
\description{
Read a parquet file into a dataframe
}
