% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_cli}
\alias{run_cli}
\title{Wrapper for `base::system2` for testing convenience}
\usage{
run_cli(command)
}
\arguments{
\item{command}{string to run as command}
}
\description{
Wrapper for `base::system2` for testing convenience
}
