% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{subset_valid_files}
\alias{subset_valid_files}
\title{Return file names only if they contain data to be used in training}
\usage{
subset_valid_files(files_list, file_type = c("daily", "rollup"), params)
}
\description{
Parse filenames to find included dates. Use different patterns if file
includes daily or rollup (multiple days) data.
}
