% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{add_issue_date}
\alias{add_issue_date}
\title{Add `issue_date` parsed from filename to daily input, if needed}
\usage{
add_issue_date(
  df,
  input_file,
  indicator,
  signal,
  lag_col = "lag",
  issued_col = "issue_date"
)
}
\arguments{
\item{df}{dataframe containing input data from a single parquet file}

\item{input_file}{path to the an individual file containing input data
in `.parquet` format.}

\item{indicator}{string specifying the name of the indicator as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{signal}{string specifying the name of the signal as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{lag_col}{string specifying name of lag field within
the input dataframe.}

\item{issued_col}{string specifying name of issue date (version) field within
the input dataframe.}
}
\description{
Add `issue_date` parsed from filename to daily input, if needed
}
