% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{add_params_for_dates}
\alias{add_params_for_dates}
\title{Add params related to date
Target is the updates made ref_lag days after the first release}
\usage{
add_params_for_dates(backfill_df, refd_col, lag_col)
}
\arguments{
\item{refd_col}{column name for the column of reference date}

\item{lag_col}{column name for the column of lag}

\item{df}{Data Frame of aggregated counts within a single location 
reported for each reference date and issue date.}
}
\description{
Add params related to date
Target is the updates made ref_lag days after the first release
}
