% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R, R/preprocessing.R
\name{add_sqrtscale}
\alias{add_sqrtscale}
\title{Add square root scale indicator}
\usage{
add_sqrtscale(train_data, test_data, max_raw, value_col)

add_sqrtscale(train_data, test_data, max_raw, value_col)
}
\arguments{
\item{train_data}{Data Frame containing training data}

\item{test_data}{Data Frame for testing}

\item{max_raw}{the maximum value in the training data at square root level}

\item{value_col}{string specifying name of value (counts) field within
the input dataframe.}
}
\description{
Add square root scale indicator

Add columns to indicate the scale of value at square root level
}
