% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{create_name_pattern}
\alias{create_name_pattern}
\title{Create pattern to match input files of a given type, signal, and geo level}
\usage{
create_name_pattern(
  indicator,
  signal,
  geo_level,
  file_type = c("daily", "rollup")
)
}
\description{
Create pattern to match input files of a given type, signal, and geo level
}
