% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior_estimation.R
\name{est_priors}
\alias{est_priors}
\title{Main function for the beta prior approach
Estimate the priors for the beta distribution based on data for 
a certain day of a week}
\usage{
est_priors(
  train_data,
  prior_test_data,
  dw,
  taus,
  covariates,
  response,
  lp_solver,
  lambda,
  start = c(0, log(10)),
  base_pseudo_denom = 1000,
  base_pseudo_num = 10
)
}
\arguments{
\item{train_data}{Data Frame containing training data}

\item{prior_test_data}{Data Frame for testing}

\item{dw}{column name to indicate which day of a week it is}

\item{taus}{numeric vector of quantiles to be predicted. Values
must be between 0 and 1.}

\item{covariates}{character vector of column names serving as the covariates for the model}

\item{response}{the column name of the response variable}

\item{lp_solver}{string specifying the lp solver to use in
Quantgen fitting. Either "glpk" or "gurobi". For faster
optimization, use Gurobi (requires separate installation
of the `gurobi` package).}

\item{lambda}{the level of lasso penalty}

\item{start}{the initialization of the the points in nlm}

\item{base_pseudo_denom}{the pseudo counts added to denominator if little data for training}

\item{base_pseudo_num}{the pseudo counts added to numerator if little data for training}
}
\description{
Main function for the beta prior approach
Estimate the priors for the beta distribution based on data for 
a certain day of a week
}
