% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{evl}
\alias{evl}
\title{Evaluation of the test results based on WIS score
The WIS score calculation is based on the weighted_interval_score function 
from the `evalcast` package from Delphi}
\usage{
evl(test_data, taus)
}
\arguments{
\item{test_data}{dataframe with a column containing the prediction results of
each requested quantile. Each row represents an update with certain
(reference_date, issue_date, location) combination.}

\item{taus}{numeric vector of quantiles to be predicted. Values
must be between 0 and 1.}
}
\description{
Evaluation of the test results based on WIS score
The WIS score calculation is based on the weighted_interval_score function 
from the `evalcast` package from Delphi
}
