% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_test_result}
\alias{export_test_result}
\title{Export the result to customized directory}
\usage{
export_test_result(
  test_data,
  coef_data,
  export_dir,
  geo_level,
  test_lag = NULL
)
}
\arguments{
\item{test_data}{test data containing prediction results}

\item{coef_data}{data frame containing the estimated coefficients}

\item{export_dir}{export directory}

\item{geo_level}{string describing geo coverage of input data. Either "state"
or "county".}

\item{test_lag}{integer number of days ago to predict for}
}
\description{
Export the result to customized directory
}
