% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_test_result}
\alias{export_test_result}
\title{Export the result to customized directory}
\usage{
export_test_result(
  test_data,
  coef_data,
  indicator,
  signal,
  geo_level,
  signal_suffix,
  lambda,
  training_end_date,
  training_start_date,
  value_type,
  export_dir
)
}
\arguments{
\item{test_data}{test data containing prediction results}

\item{coef_data}{data frame containing the estimated coefficients}

\item{indicator}{string specifying the name of the indicator as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{signal}{string specifying the name of the signal as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{geo_level}{string describing geo coverage of input data. Either "state"
or "county".}

\item{signal_suffix}{string specifying value column name
ending to be appended to standard value column names from
`params$num_col` and `params$denom_col`. Used for non-standard
value column names and when processing multiple signals from a
single input dataframe, as with `quidel`'s age buckets.}

\item{lambda}{the level of lasso penalty}

\item{training_end_date}{the latest target date and issue date included in training data}

\item{training_start_date}{the earliest target date included in training data}

\item{value_type}{string describing signal type. Either "count" or "fraction".}

\item{export_dir}{path to directory to save output to}

\item{geo}{string specifying the name of the geo region (e.g. FIPS
code for counties)}
}
\description{
Export the result to customized directory
}
