% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_test_result}
\alias{export_test_result}
\title{Export the result to customized directory}
\usage{
export_test_result(test_data, coef_data, export_dir, geo_level, test_lag)
}
\arguments{
\item{test_data}{test data with prediction result}

\item{coef_data}{data frame with the estimated coefficients}

\item{export_dir}{export directory}

\item{geo_level}{geographical level, can be county or state}

\item{test_lag}{}
}
\description{
Export the result to customized directory
}
