% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior_estimation.R
\name{frac_adj}
\alias{frac_adj}
\title{Update fraction using beta prior approach}
\usage{
frac_adj(
  train_data,
  test_data,
  prior_test_data,
  indicator,
  signal,
  geo_level,
  signal_suffix,
  lambda,
  value_type,
  geo,
  training_end_date,
  training_start_date,
  model_save_dir,
  taus,
  lp_solver,
  train_models = TRUE,
  make_predictions = TRUE
)
}
\arguments{
\item{train_data}{Data Frame containing training data}

\item{test_data}{testing data}

\item{prior_test_data}{testing data for the lag -1 model}

\item{indicator}{string specifying the name of the indicator as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{signal}{string specifying the name of the signal as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{geo_level}{string describing geo coverage of input data. Either "state"
or "county".}

\item{signal_suffix}{string specifying value column name
ending to be appended to standard value column names from
`params$num_col` and `params$denom_col`. Used for non-standard
value column names and when processing multiple signals from a
single input dataframe, as with `quidel`'s age buckets.}

\item{lambda}{the level of lasso penalty}

\item{value_type}{string describing signal type. Either "count" or "fraction".}

\item{geo}{string specifying the name of the geo region (e.g. FIPS
code for counties)}

\item{training_end_date}{the latest target date and issue date included in training data}

\item{training_start_date}{the earliest target date included in training data}

\item{model_save_dir}{directory containing trained models}

\item{taus}{numeric vector of quantiles to be predicted. Values
must be between 0 and 1.}

\item{lp_solver}{string specifying the lp solver to use in
Quantgen fitting. Either "glpk" or "gurobi". For faster
optimization, use Gurobi (requires separate installation
of the `gurobi` package).}

\item{train_models}{boolean indicating whether to train models (TRUE). If
FALSE previously trained models (stored locally) will be used instead.
Default is TRUE.}

\item{make_predictions}{boolean indicating whether to generate and save
corrections (TRUE) or not. Default is TRUE.}
}
\description{
Update fraction using beta prior approach
}
