% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooling.R
\name{main_local}
\alias{main_local}
\title{Main function to correct a single local signal}
\usage{
main_local(
  input_dir,
  export_dir,
  test_start_date,
  test_end_date,
  num_col,
  denom_col,
  value_type = c("count", "fraction"),
  training_days = TRAINING_DAYS,
  testing_window = TESTING_WINDOW,
  lambda = LAMBDA,
  ref_lag = REF_LAG,
  lp_solver = LP_SOLVER
)
}
\arguments{
\item{input_dir}{path to the directory containing input data}

\item{export_dir}{path to directory to save output to}

\item{test_start_date}{Date or string in the format "YYYY-MM-DD" to start
making predictions on}

\item{test_end_date}{Date or string in the format "YYYY-MM-DD" to stop
making predictions on}

\item{num_col}{name of numerator column in the input dataframe}

\item{denom_col}{name of denominator column in the input dataframe}

\item{value_type}{string describing signal type. Either "count" or "fraction".}

\item{training_days}{integer number of days to use for training}

\item{testing_window}{the testing window used for saving the runtime. Could
set it to be 1 if time allows}

\item{lambda}{the level of lasso penalty}

\item{ref_lag}{max lag to use for training}

\item{lp_solver}{string specifying the lp solver to use in
Quantgen fitting. Either "glpk" or "gurobi". For faster
optimization, use Gurobi (requires separate installation
of the `gurobi` package).}
}
\description{
Main function to correct a single local signal
}
