% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooling.R
\name{main_local}
\alias{main_local}
\title{Main function to correct a single local signal}
\usage{
main_local(
  data_path,
  export_dir,
  test_start_date,
  test_end_date,
  traning_days,
  testing_window,
  value_type,
  num_col,
  denom_col,
  lambda,
  ref_lag
)
}
\description{
Main function to correct a single local signal
}
