% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_training_and_testing}
\alias{model_training_and_testing}
\title{Model training and prediction using quantile regression with Lasso penalty
The quantile regression uses the quantile_lasso function from quantgen package}
\usage{
model_training_and_testing(
  train_data,
  test_data,
  taus,
  params_list,
  lp_solver,
  lambda,
  test_date
)
}
\arguments{
\item{train_data}{Data frame for training}

\item{test_data}{Data frame for testing}

\item{taus}{vector of considered quantiles}

\item{params_list}{the list of column names serving as the covariates}

\item{lp_solver}{the lp solver used in Quantgen}

\item{lambda}{the level of lasso penalty}

\item{test_date}{as.Date}
}
\description{
Model training and prediction using quantile regression with Lasso penalty
The quantile regression uses the quantile_lasso function from quantgen package
}
