% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_training_and_testing}
\alias{model_training_and_testing}
\title{Fetch model and use to generate predictions/perform corrections}
\usage{
model_training_and_testing(
  train_data,
  test_data,
  taus,
  covariates,
  lp_solver,
  lambda,
  test_lag,
  geo,
  model_save_dir,
  indicator,
  signal,
  geo_level,
  signal_suffix,
  training_end_date,
  train_models = TRUE,
  make_predictions = TRUE
)
}
\arguments{
\item{train_data}{Data Frame containing training data}

\item{test_data}{Data frame for testing}

\item{taus}{numeric vector of quantiles to be predicted. Values
must be between 0 and 1.}

\item{covariates}{character vector of column names serving as the covariates for the model}

\item{lp_solver}{string specifying the lp solver to use in
Quantgen fitting. Either "glpk" or "gurobi". For faster
optimization, use Gurobi (requires separate installation
of the `gurobi` package).}

\item{lambda}{the level of lasso penalty}

\item{test_lag}{integer number of days ago to predict for}

\item{geo}{string specifying the name of the geo region (e.g. FIPS
code for counties)}

\item{indicator}{string specifying the name of the indicator as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{signal}{string specifying the name of the signal as used in
`parquet` input data filenames. One indicator can be associated
with multiple signals.}

\item{signal_suffix}{string specifying value column name
ending to be appended to standard value column names from
`params$num_col` and `params$denom_col`. Used for non-standard
value column names and when processing multiple signals from a
single input dataframe, as with `quidel`'s age buckets.}

\item{train_models}{boolean indicating whether to train models (TRUE). If
FALSE previously trained models (stored locally) will be used instead.
Default is TRUE.}

\item{make_predictions}{boolean indicating whether to generate and save
corrections (TRUE) or not. Default is TRUE.}

\item{test_date}{Date object representing test date}

\item{value_type}{string describing signal type. Either "count" or "fraction".}
}
\description{
Fetch model and use to generate predictions/perform corrections
}
