% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg_ts}
\alias{msg_ts}
\title{Write a message to the console with the current time}
\usage{
msg_ts(text)
}
\arguments{
\item{text}{the body of the message to display}
}
\description{
Write a message to the console with the current time
}
