% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior_estimation.R
\name{ratio_adj}
\alias{ratio_adj}
\title{Update ratio using beta prior approach}
\usage{
ratio_adj(
  train_data,
  test_data,
  prior_test_data,
  taus = TAUS,
  lp_solver = LP_SOLVER
)
}
\arguments{
\item{train_data}{Data Frame containing training data}

\item{test_data}{testing data}

\item{prior_test_data}{testing data for the lag -1 model}

\item{taus}{numeric vector of quantiles to be predicted. Values
must be between 0 and 1.}

\item{lp_solver}{string specifying the lp solver to use in
Quantgen fitting. Either "glpk" or "gurobi". For faster
optimization, use Gurobi (requires separate installation
of the `gurobi` package).}
}
\description{
Update ratio using beta prior approach
}
