% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior_estimation.R
\name{ratio_adj_with_pseudo}
\alias{ratio_adj_with_pseudo}
\title{Update fraction based on the pseudo counts for numerators and denominators}
\usage{
ratio_adj_with_pseudo(data, dw, pseudo_num, pseudo_denom, num_col, denom_col)
}
\arguments{
\item{data}{Data Frame}

\item{dw}{character to indicate the day of a week. Can be NULL for all the days}

\item{pseudo_num}{the estimated counts to be added to numerators}

\item{pseudo_denom}{the estimated counts to be added to denominators}

\item{num_col}{name of numerator column in the input dataframe}

\item{denom_col}{name of denominator column in the input dataframe}
}
\description{
Update fraction based on the pseudo counts for numerators and denominators
}
