% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_backfill}
\alias{run_backfill}
\title{Get backfill-corrected estimates for a single signal + geo combination}
\usage{
run_backfill(
  df,
  value_type,
  geo_level,
  params,
  refd_col = "time_value",
  lag_col = "lag",
  signal_suffixes = c("")
)
}
\arguments{
\item{df}{dataframe of input data containing a single indicator + signal +
level of geographic coverage.}

\item{value_type}{string describing signal type of "count" and "ratio".}

\item{geo_level}{string describing geo coverage of input data. "state" or
"county". If "county" is selected, only data from the 200 most populous
counties in the US (*not* the dataset) will be used.}

\item{params}{named list containing modeling and data settings. Must include
the following elements: `ref_lag`, `testing_window`, `test_dates`,
`training_days`, `num_col`, `taus`, `lambda`, `export_dir`, `lp_solver`,
and `data_path` (input dir).}

\item{refd_col}{string containing name of reference date field within `df`.}

\item{lag_col}{string containing name of lag field within `df`.}
}
\description{
Get backfill-corrected estimates for a single signal + geo combination
}
