% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooling.R
\name{run_backfill_local}
\alias{run_backfill_local}
\title{Corrected estimates from a single local signal}
\usage{
run_backfill_local(
  df,
  export_dir,
  taus,
  test_date_list,
  test_lags,
  value_cols,
  training_days,
  testing_window,
  ref_lag,
  value_type,
  lambda
)
}
\description{
Corrected estimates from a single local signal
}
